remove Cacheable parent class, keep cache implementation flat
