+++ lib/business_calendar/holiday_determiner.rb
@@ -1,11 +1,12 @@
 require 'holidays'
-require 'business_calendar/cacheable'
 
-class BusinessCalendar::HolidayDeterminer < BusinessCalendar::Cacheable
+class BusinessCalendar::HolidayDeterminer
+  DEFAULT_TIME_TO_LIVE = 24 * 60 * 60
   attr_reader :regions, :holiday_names, :additions, :removals, :additions_only
 
   def initialize(regions, holiday_names, opts = {})
-    super(opts[:ttl])
+    ttl = opts[:ttl]
+    @time_to_live = ttl.nil? ? DEFAULT_TIME_TO_LIVE : ttl
     @regions        = regions
     @holiday_names  = holiday_names
     @additions      = opts[:additions]      || []
@@ -28,8 +29,14 @@ class BusinessCalendar::HolidayDeterminer < BusinessCalendar::Cacheable
 
   private
 
+  def should_clear_cache?
+    return false unless @time_to_live
+
+    !@last_cleared || (Time.now - @last_cleared) >= @time_to_live
+  end
+
   def clear_cache
-    super
+    @last_cleared = Time.now
     @additions_cache = nil
     @removals_cache = nil
   end
